Add AutoAPI for doing isImmutable checks.
